# backend/internal/http/static/openapi.yaml
openapi: 3.0.3
info:
  title: Noryx API
  version: 1.0.0
  description: |
    The Noryx backend API, in full: every route the platform serves is
    described here, and a test fails the build when one is not.

    Entries generated from the router carry a summary derived from the
    handler's name; the ones that needed a human sentence have one. Both are
    checked against the router, so neither can drift.

    Authentication modes:
    - `Bearer` token (OIDC / Keycloak)
    - bootstrap compatibility header `X-Noryx-User` (lab mode)
servers:
  - url: /
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    xNoryxUser:
      type: apiKey
      in: header
      name: X-Noryx-User
  schemas:
...
    JobDataset:
      type: object
      description: >-
        One mount as it was when the job started. The name and bucket travel
        with the identifier because a dataset renamed, moved or deleted since
        must still mean something to whoever reads this.
      properties:
        id: { type: string }
        name: { type: string }
        bucket: { type: string }
        prefix: { type: string }
        readOnly:
          type: boolean
          description: Whether the job could have changed what it read.
...
    BuildLogsResponse:
      type: object
      properties:
        buildId: { type: string }
        jobName: { type: string }
        projectId: { type: string }
        status: { type: string }
        podName: { type: string }
        logs: { type: string }
        pending: { type: boolean }
        unavailable: { type: string }
...
    App:
      type: object
      properties:
        id:
          type: string
    AppListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/App'
    Job:
      type: object
      properties:
        id:
          type: string
        datasets:
          type: array
          description: >-
            What this job had mounted when it ran, recorded on the job rather
            than left to be reconstructed from the project: a project's
            attachments change and a result does not. Empty for runs that
            predate this, which is different from a run that read nothing.
          items:
            $ref: '#/components/schemas/JobDataset'
    JobListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Job'
    Datasource:
      type: object
      properties:
        id:
          type: string
    DatasourceListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Datasource'
    Ontology:
      type: object
      properties:
        id:
          type: string
    OntologyListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ontology'
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Not enough permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/v1/apps/{appID}/logs:
    get:
      tags: [Workloads]
      summary: Read a workload's recent output
      operationId: getAppLogs
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: appID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/datasources/{datasourceID}/logs:
    get:
      tags: [Datasets]
      summary: Get Data Service Logs
      operationId: getDataServiceLogs
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: datasourceID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/jobs/{jobID}/logs:
    get:
      tags: [Workloads]
      summary: Get Job Logs
      operationId: getJobLogs
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: jobID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/projects/{projectID}/ontology:
    get:
      tags: [Projects]
      summary: Get Project Ontology
      operationId: getProjectOntology
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: projectID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/builds/{buildID}/logs:
    get:
      tags: [Workloads]
      summary: What a build printed, so a failed one can be read
      operationId: getBuildLogs
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: buildID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildLogsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/ontologies/{ontologyID}/freshness:
    get:
      tags: [Datasets]
      summary: Get Ontology Freshness
      operationId: getOntologyFreshness
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: ontologyID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OntologyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/apis/{apiID}/logs:
    get:
      tags: [Platform]
      summary: Read a workload's recent output
      operationId: getAppLogs
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: apiID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/dashboards/{dashboardID}/logs:
    get:
      tags: [Workloads]
      summary: Read a workload's recent output
      operationId: getAppLogs
      security:
        - bearerAuth: []
        - xNoryxUser: []
      parameters:
        - name: dashboardID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...